refactor: déplacer composante à fins de réutilisation
